Created switchSlide function in ApiController
This function is called whenever the professor switches to another slide during his presentation.
This function looks if there is question in DB for the current slide.
If yes, than push the question to all students which have subscribed for this lecture.
Still some TODOs to be done at this commit ....